- Updated so that list of certified sites are fetched from a google sheet instead of a json filed stored in our repo
